# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Torneo Socorro Senior Master Plus 40 - Tournament Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partidos/{id}/planilla:
    get:
      operationId: getPlanilla
      tags: [planilla]
      summary: Match roster (planilla)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Roster for both teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Planilla"
        "404":
          description: Not found
    put:
      operationId: savePlanilla
      tags: [planilla]
      summary: Save the full match roster and recompute the score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePlanillaInput"
      responses:
        "204":
          description: Saved
        "409":
          description: Match is a walkover and has no roster

  # ── GOLES ──────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    PlanillaJugador:
      type: object
      required: [jugadorId, jugadorNombre, equipoId, jugo, titular, goles, amarillas, rojas]
      properties:
        jugadorId:
          type: number
        jugadorNombre:
          type: string
        equipoId:
          type: number
        nCarnet:
          type: ["number", "null"]
        jugo:
          type: boolean
        dorsal:
          type: ["number", "null"]
        titular:
          type: boolean
        goles:
          type: number
        amarillas:
          type: number
        rojas:
          type: number
        fechasSancion:
          type: number
        # Fechas de sanción que el jugador todavía NO había cumplido cuando
        # se jugó este partido. Mayor que cero = no se puede alinear.
        fechasPendientes:
          type: number
    Planilla:
      type: object
      required: [partidoId, localId, visitanteId, jugadores]
      properties:
        partidoId:
          type: number
        localId:
          type: number
        visitanteId:
          type: number
        arbitroId:
          type: ["number", "null"]
        arbitroNombre:
          type: ["string", "null"]
        mesa:
          type: ["string", "null"]
        jugadores:
          type: array
          items:
            $ref: "#/components/schemas/PlanillaJugador"
    SavePlanillaInput:
      type: object
      required: [jugadores]
      properties:
        arbitroId:
          type: ["number", "null"]
        mesa:
          type: string
        # Ya no se manda valorAmarilla/valorRoja desde acá: el valor de cada
        # tarjeta nueva se asigna solo, tomándolo de /ajustes en el momento
        # de guardar la planilla (ver routes/planilla.ts).
        jugadores:
          type: array
          items:
            type: object
            required: [jugadorId, jugo]
            properties:
              jugadorId:
                type: number
              jugo:
                type: boolean
              dorsal:
                type: ["number", "null"]
              titular:
                type: boolean
              goles:
                type: number
                minimum: 0
              amarillas:
                type: number
                minimum: 0
                maximum: 2
              rojas:
                type: number
                minimum: 0
                maximum: 1
              fechasSancion:
                type: number
                minimum: 0
                maximum: 20

    # ── GOLES ────────────────────────────────────────────────────────────────
...